Introduce function_id for instrumented functions
Previously the client code used absolute function address to identify
the instrumented function in timers and in some other places in UI.

Since using the address assumes that the function address in memory is
fixed (which might not be true) this has been replaced with function_id
for instrumented functions.

Changed code relying on function absolute address where it is not needed
and replaced it with function_id.

Test: run unit tests
Test: hook functions run capture, create iterators and frame tracks
Bug: http://b/177525225
Bug: http://b/175865913

+++ src/OrbitGl/CallStackDataView.cpp
@@ -161,7 +161,7 @@ void CallStackDataView::DoFilter() {
     return;
   }
 
-  std::vector<uint32_t> indices;
+  std::vector<uint64_t> indices;
   std::vector<std::string> tokens = absl::StrSplit(ToLower(filter_), ' ');
 
   for (size_t i = 0; i < callstack_.GetFramesCount(); ++i) {
